contract-checker/archief/contract-check/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
anthropic>=0.40.0

# Database (psycopg3 for Python 3.13+ compatibility)
psycopg[binary]>=3.2.0
SQLAlchemy>=2.0.23

# Supabase for LLM configuration
supabase>=2.3.0

# Additional LLM providers
openai>=1.12.0
requests>=2.31.0

# Document processing
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.2.0
pdfplumber>=0.10.0

# Utilities
python-dateutil>=2.8.2

# Frontend
streamlit>=1.37.0
plotly>=5.18.0

# Type stubs (for IDE type checking)
types-openpyxl>=3.1.0
lxml-stubs>=0.5.0
